Fix batch import input/return data type casting
This commit makes the following changes to fix issues with the data sent
and returned from the batch import function when run by the app's database
module:

- Add explicit NUMERIC cast to tax rate calculations in view
- Add COALESCE guards for NULL safety in return values

I've also made two changes to simplify the query itself:

- Replace subquery with INNER JOIN for tax year lookup
- Calculate estimated tax directly in the summary instead of using the view

type: fix
scope: database

+++ src/FoxyBalance.Migrations/Migrations/Migration_001_InitialPostgreSQLSchema.fs
@@ -91,7 +91,7 @@ type Migration_001_InitialPostgreSQLSchema() =
                 ir.platformfee,
                 ir.processingfee,
                 ir.netshare,
-                (ir.netshare * ty.taxrate / 100) AS estimatedtax,
+                (ir.netshare * ty.taxrate::NUMERIC / 100) AS estimatedtax,
                 ir.ignored
             FROM foxybalance_incomerecords ir
             INNER JOIN foxybalance_taxyears ty ON ir.taxyearid = ty.id
